flask/requirements.txt
Flask
gensim
pyyaml
janome
keras
tensorflow
